Golang initial check-in
First checkin. Still needs a lot more testing, but other people asked to see it.

This is my first significant project in Go, and has not been reviewed by anyone experienced in the language. Since I really like the language, I'm hoping to come back to this in a year or so and note my Lessons Learned. (My other projects here in hello-smtp started out as "my first program," but were written years ago and have evolved quite a bit. This Golang implementation is the first time I've made my first revision public.)